pyproject.toml
[tool.poetry]
name = "tongle"
version = "0.1.0"
description = "A translation discord bot."
authors = ["Nathan Solis <66754842+colderinit@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
deep-translator = "^1.8.3"
python-dotenv = "^0.21.0"
discord-py-interactions = "^4.3.4"
interactions-wait-for = "^1.0.6"
pymongo = {extras = ["srv"], version = "^4.3.2"}

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.5"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
